Added Virtual Properties and Template trait
VirtualProperties makes it easier to use virtual properties
in classes by creating methods with a special prefix,
they will then be available as virtual properties.

TemplateAttributes is a trait for making any class
properties available as attributes in the eZ publish template
engine. It is best paired with the VirtualProperties trait.

See the README or docblock for more details.